Fixing circular import problem
