+++ src/csvcubed/inspect/browsertable.py
@@ -1,5 +1,6 @@
-from dataclasses import dataclass
-from typing import OrderedDict
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import List, OrderedDict, Union
 
 import uritemplate
 
@@ -12,7 +13,6 @@ from csvcubed.inspect.browsercolumns import (
     StandardShapeObservationsColumn,
     SuppressedColumn,
     UnitsColumn,
-    _map_to_data_cube_column,
 )
 from csvcubed.inspect.browsers import MetadataBrowser, TableBrowser
 from csvcubed.inspect.lazyfuncdescriptor import lazy_func_field
@@ -20,7 +20,11 @@ from csvcubed.models.cube.cube_shape import CubeShape
 from csvcubed.models.sparqlresults import CatalogMetadataResult
 from csvcubed.utils.iterables import single
 from csvcubed.utils.qb.components import EndUserColumnType
+from csvcubed.utils.sparql_handler.code_list_inspector import CodeListInspector
 from csvcubed.utils.sparql_handler.column_component_info import ColumnComponentInfo
+from csvcubed.utils.sparql_handler.csvw_inspector import CsvWInspector
+from csvcubed.utils.sparql_handler.data_cube_inspector import DataCubeInspector
+from csvcubed.utils.tableschema import CsvWRdfManager
 
 
 @dataclass(frozen=True)
@@ -128,3 +132,44 @@ def _map_to_data_cube_column(
     raise ValueError(
         f"Unmatched column type {info.column_type} with column title {info.column_definition.title}"
     )
+
+
+@dataclass
+class CsvWBrowser:
+    primary_csvw: Union[str, Path]
+    _csvw_inspector: CsvWInspector = field(init=False, repr=False)
+    _data_cube_inspector: DataCubeInspector = field(init=False, repr=False)
+    _code_list_inspector: CodeListInspector = field(init=False, repr=False)
+
+    def _get_tables(self) -> List[TableBrowser]:
+        cube_tables = [
+            DataCubeTable(
+                csv_url=t.csv_url,
+                data_cube_inspector=self._data_cube_inspector,
+                code_list_inspector=self._code_list_inspector,
+            )
+            for t in self._data_cube_inspector._cube_table_identifiers.values()
+        ]
+        code_list_tables = [
+            CodeListTable(
+                csv_url=t.csv_url,
+                data_cube_inspector=self._data_cube_inspector,
+                code_list_inspector=self._code_list_inspector,
+            )
+            for t in self._code_list_inspector._code_list_table_identifiers
+        ]
+
+        return [*cube_tables, *code_list_tables]
+
+    tables: List[TableBrowser] = lazy_func_field(_get_tables)
+
+    def __post_init__(self):
+        csvw_path = (
+            self.primary_csvw
+            if isinstance(self.primary_csvw, Path)
+            else Path(self.primary_csvw)
+        )
+        csvw_rdf_manager = CsvWRdfManager(csvw_path.expanduser())
+        self._csvw_inspector = csvw_rdf_manager.csvw_inspector
+        self._data_cube_inspector = DataCubeInspector(self._csvw_inspector)
+        self._code_list_inspector = CodeListInspector(self._csvw_inspector)
